Code Review: Updating code base on reviews and refactoring test case
