+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/util/ServiceApplicationRefusalHelper.java
@@ -2,17 +2,19 @@ package uk.gov.hmcts.reform.divorce.orchestration.util;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CaseDetails;
-import uk.gov.hmcts.reform.divorce.orchestration.domain.model.document.ApplicationServiceTypes;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.document.ServiceRefusalDecision;
 
 import java.util.Map;
 
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields.SERVICE_APPLICATION_GRANTED;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields.SERVICE_APPLICATION_REFUSAL_REASON;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields.SERVICE_APPLICATION_TYPE;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdStates.AWAITING_SERVICE_CONSIDERATION;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.CASE_ID_JSON_KEY;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.YES_VALUE;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.document.ApplicationServiceTypes.DEEMED;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.document.ApplicationServiceTypes.DISPENSED;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.document.ServiceRefusalDecision.DRAFT;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.document.ServiceRefusalDecision.FINAL;
 import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.helper.ExtractorHelper.getMandatoryStringValue;
@@ -20,7 +22,7 @@ import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.helpe
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class ServiceApplicationRefusalHelper {
     public static String getServiceApplicationRefusalReason(Map<String, Object> caseData) {
-        return getMandatoryStringValue(caseData, CcdFields.SERVICE_APPLICATION_REFUSAL_REASON);
+        return getMandatoryStringValue(caseData, SERVICE_APPLICATION_REFUSAL_REASON);
     }
 
     public static String getServiceApplicationGranted(Map<String, Object> caseData) {
@@ -28,7 +30,7 @@ public class ServiceApplicationRefusalHelper {
     }
 
     public static String getServiceApplicationType(Map<String, Object> caseData) {
-        return getMandatoryStringValue(caseData, CcdFields.SERVICE_APPLICATION_TYPE);
+        return getMandatoryStringValue(caseData, SERVICE_APPLICATION_TYPE);
     }
 
     public static String getCaseReference(Map<String, Object> caseData) {
@@ -44,18 +46,18 @@ public class ServiceApplicationRefusalHelper {
     }
 
     public static boolean isDispensedApplication(String applicationType) {
-        return ApplicationServiceTypes.DISPENSED.equals(applicationType);
+        return DISPENSED.equalsIgnoreCase(applicationType);
     }
 
     public static boolean isDeemedApplication(String applicationType) {
-        return ApplicationServiceTypes.DEEMED.equals(applicationType);
+        return DEEMED.equalsIgnoreCase(applicationType);
     }
 
     public static boolean isAwaitingServiceConsideration(CaseDetails caseDetails) {
-        return AWAITING_SERVICE_CONSIDERATION.equals(caseDetails.getState());
+        return AWAITING_SERVICE_CONSIDERATION.equalsIgnoreCase(caseDetails.getState());
     }
 
     public static boolean isServiceApplicationGranted(Map<String, Object> caseData) {
-        return YES_VALUE.equals(getMandatoryStringValue(caseData, SERVICE_APPLICATION_GRANTED));
+        return YES_VALUE.equalsIgnoreCase(getMandatoryStringValue(caseData, SERVICE_APPLICATION_GRANTED));
     }
 }
