+++ src/test/java/uk/gov/hmcts/reform/divorce/orchestration/tasks/servicejourney/DeemedServiceRefusalOrderTaskTest.java
@@ -6,6 +6,10 @@ import org.mockito.InjectMocks;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.task.TaskException;
 
+import java.util.Map;
+
+import static org.junit.Assert.assertNotNull;
+
 @RunWith(MockitoJUnitRunner.class)
 public class DeemedServiceRefusalOrderTaskTest extends ServiceRefusalOrderGenerationTaskTest {
 
@@ -19,6 +23,7 @@ public class DeemedServiceRefusalOrderTaskTest extends ServiceRefusalOrderGenera
 
     @Test
     public void testExecuteShouldGenerateAFile() throws TaskException {
-        executeShouldGenerateAFile();
+        Map<String, Object> returnedCaseData = executeShouldGenerateAFile();
+        assertNotNull(returnedCaseData);
     }
 }
